rework allocator to now abstract common allocations
sp_alloc_init is now used to configure the allocator. This must be
called in a thread-safe manner; that is, the alloc module does not
enforce thread safety at the moment.

sp_alloc_debug is the old debugging allocator, but it has been reworked
to emit better errors.

sp_alloc_system is the default allocator that uses realloc/free.

sp_realloc and sp_free now require taking the current allocation size
as an argument. The underlying allocator does not need to use this
information, but it enables a much more performant allocator down the
road.