rework directory tree walker to include stat tracking
This will only stat the file if the type is unknown or an explicit stat
is requested. This will not follow symlinks unless the `sp_dir_follow`
is called. This will call lstat on the symlink, thus updating the type
and allowing the walker to descend. It is not currently checking for
recursive links, so beware.